Basic Save, a working Pause Menu and small adjustment to Start Menu
For testing purposes and for further use I've made a pause menu for saving and loading a save. For now all that is saved is the name of the level the player is in and the player's transform (used to access the player's position in the level). Also fixed an issue with the start menu where the input mode now stays on game only with the mouse cursor hidden whenever the player presses the start button